+++ packages/api/src/utils/building-placeholder.ts
@@ -5,9 +5,11 @@ import {
 } from '@pillage-first/game-assets/utils/buildings';
 import type { Building } from '@pillage-first/types/models/building';
 import type { BuildingField } from '@pillage-first/types/models/building-field';
+import { tribeSchema } from '@pillage-first/types/models/tribe';
 import type { Village } from '@pillage-first/types/models/village';
 import type { DbFacade } from '@pillage-first/utils/facades/database';
 import { updatePopulationEffectQuery } from '../queries/effect-queries';
+import { selectVillageTribeQuery } from '../queries/village-queries';
 import { demolishBuilding } from './village';
 
 export const createBuildingPlaceholder = (
@@ -30,7 +32,13 @@ export const createBuildingPlaceholder = (
     },
   });
 
-  const { effects } = getBuildingDefinition(buildingId);
+  const tribe = database.selectValue({
+    sql: selectVillageTribeQuery,
+    bind: { $village_id: villageId },
+    schema: tribeSchema,
+  })!;
+
+  const effects = getBuildingDefinition(buildingId).effects(tribe);
 
   database.exec({
     sql: `